requirements.txt
python-telegram-bot==13.7
requests==2.31.0
openai==1.2.3
